+++ deepfence_frontend/apps/dashboard/src/components/forms/SearchableContainerList.tsx
@@ -1,15 +1,21 @@
 import { useSuspenseInfiniteQuery } from '@suspensive/react-query';
-import { debounce } from 'lodash-es';
 import { Suspense, useEffect, useMemo, useState } from 'react';
-import { CircleSpinner, Combobox, ComboboxOption } from 'ui-components';
+import {
+  CircleSpinner,
+  ComboboxV2Content,
+  ComboboxV2Item,
+  ComboboxV2Provider,
+  ComboboxV2TriggerButton,
+  ComboboxV2TriggerInput,
+} from 'ui-components';
 
 import { queries } from '@/queries';
 import { ScanTypeEnum } from '@/types/common';
+import { useDebouncedValue } from '@/utils/useDebouncedValue';
 
-export type Props = {
+export interface Props {
   scanType: ScanTypeEnum | 'none';
   onChange?: (value: string[]) => void;
-  onClearAll?: () => void;
   defaultSelectedContainers?: string[];
   valueKey?: 'nodeId' | 'nodeName';
   active?: boolean;
@@ -17,13 +23,12 @@ export type Props = {
   isScannedForVulnerabilities?: boolean;
   isScannedForSecrets?: boolean;
   isScannedForMalware?: boolean;
-};
+}
 const fieldName = 'containerFilter';
 const PAGE_SIZE = 15;
 const SearchableContainer = ({
   scanType,
   onChange,
-  onClearAll,
   defaultSelectedContainers,
   valueKey = 'nodeId',
   active = true,
@@ -33,6 +38,7 @@ const SearchableContainer = ({
   isScannedForMalware,
 }: Props) => {
   const [searchText, setSearchText] = useState('');
+  const debouncedSearchText = useDebouncedValue(searchText, 500);
 
   const [selectedContainers, setSelectedContainers] = useState<string[]>(
     defaultSelectedContainers ?? [],
@@ -51,7 +57,7 @@ const SearchableContainer = ({
       ...queries.search.containers({
         scanType,
         size: PAGE_SIZE,
-        searchText,
+        searchText: debouncedSearchText,
         active,
         isScannedForVulnerabilities,
         isScannedForSecrets,
@@ -63,6 +69,7 @@ const SearchableContainer = ({
       }),
       keepPreviousData: true,
       getNextPageParam: (lastPage, allPages) => {
+        if (lastPage.containers.length < PAGE_SIZE) return null;
         return allPages.length * PAGE_SIZE;
       },
       getPreviousPageParam: (firstPage, allPages) => {
@@ -71,55 +78,59 @@ const SearchableContainer = ({
       },
     });
 
-  const searchContainer = debounce((query) => {
-    setSearchText(query);
-  }, 1000);
-
   const onEndReached = () => {
     if (hasNextPage) fetchNextPage();
   };
 
   return (
     <>
-      <Combobox
-        startIcon={
-          isFetchingNextPage ? <CircleSpinner size="sm" className="w-3 h-3" /> : undefined
-        }
-        name={fieldName}
-        triggerVariant={triggerVariant || 'button'}
-        label={isSelectVariantType ? 'Container' : undefined}
-        getDisplayValue={() =>
-          isSelectVariantType && selectedContainers.length > 0
-            ? `${selectedContainers.length} selected`
-            : null
-        }
-        placeholder="Select container"
-        multiple
-        value={selectedContainers}
-        onChange={(values) => {
-          setSelectedContainers(values);
-          onChange?.(values);
+      <ComboboxV2Provider
+        selectedValue={selectedContainers}
+        setSelectedValue={(values) => {
+          setSelectedContainers(values as string[]);
+          onChange?.(values as string[]);
         }}
-        onQueryChange={searchContainer}
-        clearAllElement="Clear"
-        onClearAll={onClearAll}
-        onEndReached={onEndReached}
+        value={searchText}
+        setValue={setSearchText}
+        defaultSelectedValue={defaultSelectedContainers}
+        name={fieldName}
+        loading={isFetchingNextPage}
       >
-        {data?.pages
-          .flatMap((page) => {
-            return page.containers;
-          })
-          .map((container, index) => {
-            return (
-              <ComboboxOption
-                key={`${container.nodeId}-${index}`}
-                value={container[valueKey]}
-              >
-                {container.nodeName}
-              </ComboboxOption>
-            );
-          })}
-      </Combobox>
+        {isSelectVariantType ? (
+          <ComboboxV2TriggerInput
+            getDisplayValue={() =>
+              selectedContainers.length > 0
+                ? `${selectedContainers.length} selected`
+                : null
+            }
+            placeholder="Select container"
+            label="Container"
+          />
+        ) : (
+          <ComboboxV2TriggerButton>Select container</ComboboxV2TriggerButton>
+        )}
+        <ComboboxV2Content
+          width={isSelectVariantType ? 'anchor' : 'fixed'}
+          clearButtonContent="Clear"
+          onEndReached={onEndReached}
+          searchPlaceholder="Search"
+        >
+          {data?.pages
+            .flatMap((page) => {
+              return page.containers;
+            })
+            .map((container, index) => {
+              return (
+                <ComboboxV2Item
+                  key={`${container.nodeId}-${index}`}
+                  value={container[valueKey]}
+                >
+                  {container.nodeName}
+                </ComboboxV2Item>
+              );
+            })}
+        </ComboboxV2Content>
+      </ComboboxV2Provider>
     </>
   );
 };
@@ -134,18 +145,24 @@ export const SearchableContainerList = (props: Props) => {
     <Suspense
       fallback={
         <>
-          <Combobox
+          <ComboboxV2Provider
+            defaultSelectedValue={defaultSelectedContainers}
             name={fieldName}
-            value={defaultSelectedContainers}
-            label={isSelectVariantType ? 'Container' : undefined}
-            triggerVariant={triggerVariant}
-            startIcon={<CircleSpinner size="sm" className="w-3 h-3" />}
-            placeholder="Select container"
-            multiple
-            onQueryChange={() => {
-              // no operation
-            }}
-          />
+          >
+            {isSelectVariantType ? (
+              <ComboboxV2TriggerInput
+                placeholder="Select container"
+                label="Container"
+                startIcon={<CircleSpinner size="sm" className="w-3 h-3" />}
+              />
+            ) : (
+              <ComboboxV2TriggerButton
+                startIcon={<CircleSpinner size="sm" className="w-3 h-3" />}
+              >
+                Select container
+              </ComboboxV2TriggerButton>
+            )}
+          </ComboboxV2Provider>
         </>
       }
     >
