+++ deepfence_frontend/apps/dashboard/src/features/integrations/components/ReportsTable.tsx
@@ -32,7 +32,7 @@ import {
 } from '@/api/generated';
 import { FilterBadge } from '@/components/filters/FilterBadge';
 import { SearchableClusterList } from '@/components/forms/SearchableClusterList';
-import { SearchableContainerListV2 } from '@/components/forms/SearchableContainerListV2';
+import { SearchableContainerList } from '@/components/forms/SearchableContainerList';
 import { SearchableHostList } from '@/components/forms/SearchableHostList';
 import { SearchableImageList } from '@/components/forms/SearchableImageList';
 import { EllipsisIcon } from '@/components/icons/common/Ellipsis';
@@ -583,7 +583,7 @@ export const ReportFilters = () => {
             });
           }}
         />
-        <SearchableContainerListV2
+        <SearchableContainerList
           scanType={'none'}
           triggerVariant="button"
           defaultSelectedContainers={searchParams.getAll('container')}
