+++ deepfence_frontend/apps/dashboard/src/features/malwares/pages/MalwareScans.tsx
@@ -41,7 +41,7 @@ import {
 import { DFLink } from '@/components/DFLink';
 import { FilterBadge } from '@/components/filters/FilterBadge';
 import { SearchableClusterList } from '@/components/forms/SearchableClusterList';
-import { SearchableContainerListV2 } from '@/components/forms/SearchableContainerListV2';
+import { SearchableContainerList } from '@/components/forms/SearchableContainerList';
 import { SearchableHostList } from '@/components/forms/SearchableHostList';
 import { SearchableImageList } from '@/components/forms/SearchableImageList';
 import { SearchableRegistryAccountList } from '@/components/forms/SearchableRegistryAccountList';
@@ -528,7 +528,7 @@ const Filters = () => {
             });
           }}
         />
-        <SearchableContainerListV2
+        <SearchableContainerList
           scanType={ScanTypeEnum.MalwareScan}
           defaultSelectedContainers={searchParams.getAll('containers')}
           isScannedForMalware
