+++ deepfence_frontend/apps/dashboard/src/features/vulnerabilities/pages/RuntimeBom.tsx
@@ -20,7 +20,7 @@ import {
 import { DFLink } from '@/components/DFLink';
 import { FilterBadge } from '@/components/filters/FilterBadge';
 import { SearchableClusterList } from '@/components/forms/SearchableClusterList';
-import { SearchableContainerListV2 } from '@/components/forms/SearchableContainerListV2';
+import { SearchableContainerList } from '@/components/forms/SearchableContainerList';
 import { SearchableHostList } from '@/components/forms/SearchableHostList';
 import { SearchableImageList } from '@/components/forms/SearchableImageList';
 import { FileLineIcon } from '@/components/icons/common/FileLine';
@@ -247,7 +247,7 @@ const Filters = () => {
             });
           }}
         />
-        <SearchableContainerListV2
+        <SearchableContainerList
           scanType={ScanTypeEnum.VulnerabilityScan}
           defaultSelectedContainers={searchParams.getAll('containers')}
           onChange={(value) => {
